refactoring

+++ src/main/java/core/IntegralSolver.java
@@ -6,5 +6,5 @@ import util.function.ExtendedFunction;
  * @author Arthur Kupriyanov on 18.03.2020
  */
 public interface IntegralSolver {
-    double solve(ExtendedFunction function, int top, int bottom, double partition);
+    double solve(ExtendedFunction function, double top, double bottom, double partition);
 }
